Split extension into PagesExtension and PagesReactExtension, refactor PageBreakPlugin to PageBreakExtension
