+++ packages/lexical-playground/src/App.tsx
@@ -60,7 +60,8 @@ import {EmojisExtension} from './plugins/EmojisExtension';
 import {ImagesExtension} from './plugins/ImagesExtension';
 import {PlaygroundMarkdownShortcutsExtension} from './plugins/MarkdownShortcutsExtension';
 import {MaxLengthExtension} from './plugins/MaxLengthPlugin';
-import {PagesExtension} from './plugins/PagesExtension';
+import {PageBreakExtension} from './plugins/PageBreakExtension';
+import {PagesReactExtension} from './plugins/PagesReactExtension';
 import PasteLogPlugin from './plugins/PasteLogPlugin';
 import TestRecorderPlugin from './plugins/TestRecorderPlugin';
 import TypingPerfPlugin from './plugins/TypingPerfPlugin';
@@ -161,6 +162,8 @@ const PlaygroundRichTextExtension = defineExtension({
     configExtension(ListExtension, {shouldPreserveNumbering: false}),
     CheckListExtension,
     PlaygroundMarkdownShortcutsExtension,
+    PageBreakExtension,
+    PagesReactExtension,
   ],
   name: '@lexical/playground/RichText',
 });
@@ -175,7 +178,6 @@ const AppExtension = defineExtension({
     HashtagExtension,
     DateTimeExtension,
     MaxLengthExtension,
-    PagesExtension,
     DragDropPasteExtension,
     EmojisExtension,
     configExtension(LinkExtension, {validateUrl}),
