+++ packages/lexical-playground/src/plugins/PagesExtension/index.ts
@@ -5,6 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
+export {
+  PageSetupDropdownComponent,
+  type PageSetupDropdownProps,
+} from '../PagesReactExtension/PageSetupDropdown';
 export {DEFAULT_PAGE_SETUP, PAGE_SIZES} from './constants';
 export {
   $createPageContentNode,
@@ -13,9 +17,5 @@ export {
 } from './PageContentNode';
 export {$createPageNode, $isPageNode, PageNode} from './PageNode';
 export {$getPageSetup, $setPageSetup, pageSetupState} from './pageSetup';
-export {
-  PageSetupDropdownComponent,
-  type PageSetupDropdownProps,
-} from './PageSetupDropdown';
 export {type PagesConfig, PagesExtension} from './PagesExtension';
 export type {Orientation, PageSetup, PageSize} from './types';
